+++ src/controllers/OnSearchController.js
@@ -1,14 +1,14 @@
 import LoggingService from '../services/LoggingService';
 import SearchService from '../services/SearchService';
-import authVerifier from '../utilities/AuthHeaderVerifier';
 import genericResponse from './GenericResponse';
+import authVerifier from '../utilities/SignVerify/AuthHeaderVerifier';
 
 const BecknGateway = 'BG';
 const onSearch = (req, res) => {
   const logger = LoggingService.getLogger('OnSearchController');
   logger.debug(`on_search called with ${JSON.stringify(req.body)}`);
 
-  authVerifier.authorize(req, 'BG').then(() => {
+  authVerifier.authorize(req, BecknGateway).then(() => {
     SearchService.storeSearchResult(req.body);
     genericResponse.sendAcknowledgement(res);
   }).catch((err) => {
